requirements.txt
# ── Web ───────────────────────────────────────────────────────────────────────
fastapi==0.115.5
uvicorn[standard]==0.32.1
python-multipart==0.0.12

# ── Celery & Redis ────────────────────────────────────────────────────────────
celery[redis]==5.4.0
redis==5.2.0
kombu==5.4.2

# ── LangGraph / LangChain ─────────────────────────────────────────────────────
langgraph==0.2.50
langchain==0.3.9
langchain-openai==0.2.10
langchain-community==0.3.9
langchain-text-splitters==0.3.2

# ── LLM ───────────────────────────────────────────────────────────────────────
openai==1.54.4
tiktoken==0.8.0

# ── PDF Parsing ───────────────────────────────────────────────────────────────
pdfplumber==0.11.4
unstructured[pdf]==0.16.10
Pillow==11.0.0

# ── Word / HTML Parsing ────────────────────────────────────────────────────────
python-docx==1.1.2
beautifulsoup4==4.12.3
lxml==5.3.0

# ── Database ──────────────────────────────────────────────────────────────────
sqlalchemy[asyncio]==2.0.36
alembic==1.14.0
asyncpg==0.30.0
psycopg2-binary==2.9.10
pgvector==0.3.5

# ── Observability ─────────────────────────────────────────────────────────────
langfuse==2.57.0

# ── Reranking ─────────────────────────────────────────────────────────────────
# FlagEmbedding pulls torch (~2 GB). Install separately when needed:
#   pip install FlagEmbedding
# Without it the reranker falls back to RRF score order (no reranking).
FlagEmbedding==1.2.11

# ── Config / Utils ────────────────────────────────────────────────────────────
pydantic-settings==2.6.1
python-dotenv==1.0.1
httpx==0.27.2
aiofiles==24.1.0
numpy==1.26.4
